refactor: 추천 로직 분리
