+++ src/_hooks/fetcher/news/useGetNewInfo.ts
@@ -1,10 +1,10 @@
 import getNewsItemInfo from "@/services/news/GetNewsItemInfo";
 import { useQuery } from "@tanstack/react-query";
 
-const useGetNewsInfoData = (id: string, token: string) => {
+const useGetNewsInfoData = (id: string) => {
   return useQuery({
     queryKey: ["getNewsInfo", id],
-    queryFn: () => getNewsItemInfo({ id, token }),
+    queryFn: () => getNewsItemInfo({ id }),
     retry: 1,
     staleTime: 1000 * 60 * 5,
     gcTime: 1000 * 60 * 30,
